Create separate files for mappers JC-644
